Reorganizing classes
